hp5/testing/requirements.txt
fastapi
uvicorn
opencv-python
pyserial

